requirements.txt
 
Django==5.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gunicorn==21.2.0